refactor(routes): organize backend routes into feature modules
Move route handlers from src/app.rs into feature-oriented modules:
- Add src/routes/error.rs for shared error_response helper
- Add src/routes/recordings.rs for recording handlers and tests
- Add src/routes/watch.rs for watch ticket and page rendering
- Add src/routes/auth.rs for auth route builder
- Add src/routes/twitch.rs for Twitch OAuth route builder
- Add src/routes/chat.rs for chat route builder
- Add src/routes/stream.rs for stream proxy route builder

Update src/routes/mod.rs to export new modules.
Update src/routes/channels.rs to use shared error helper.
Refactor src/app.rs to focus on state construction and router composition.

Behavior preserved: all API paths, DTOs, auth behavior unchanged.
All 57 tests pass. No new dependencies. No frontend changes.
