+++ src/app.rs
@@ -1,37 +1,23 @@
 use std::path::PathBuf;
 
-use axum::{
-    Json, Router,
-    body::{Body, Bytes},
-    extract::{Path, Query, State},
-    http::StatusCode,
-    http::{HeaderMap, HeaderValue, header},
-    middleware,
-    response::{Html, IntoResponse, Response},
-    routing::{delete, get, post},
-};
-use futures_util::stream;
-use serde::{Deserialize, Serialize};
+use axum::Router;
 use tower_http::services::{ServeDir, ServeFile};
 
 use crate::{
-    auth::{self, WebAuthConfig},
-    channel_catalog::{CatalogChannel, ChannelCatalogService},
-    channels, chat,
-    config::AppConfig,
-    error::AppError,
-    live_status::LiveStatusService,
-    playback::{PlaybackTicketError, PlaybackTicketService},
-    prewarm::PrewarmCoordinator,
-    recording::{
-        ActiveRecording, RecordingBucket, RecordingError, RecordingMode, RecordingProcessingConfig,
-        RecordingService,
-    },
-    recording_rules::{self, RecordingRule},
-    recording_scheduler::RecordingScheduler,
-    routes::{self, APP_VERSION, ChannelState, LiveStatusState},
-    stream_proxy, twitch_auth, youtube, youtube_channels,
+    auth::WebAuthConfig, channel_catalog::ChannelCatalogService, channels, chat, config::AppConfig,
+    error::AppError, live_status::LiveStatusService, playback::PlaybackTicketService,
+    prewarm::PrewarmCoordinator, recording::RecordingProcessingConfig, recording::RecordingService,
+    recording_scheduler::RecordingScheduler, routes, stream_proxy, twitch_auth, youtube,
 };
+
+/// State shared between channel list and watch routes.
+#[derive(Debug, Clone)]
+pub struct ProtectedState {
+    pub auth: WebAuthConfig,
+    pub playback: PlaybackTicketService,
+    pub stream: stream_proxy::StreamSessionService,
+    pub catalog: ChannelCatalogService,
+}
 
 pub fn build_router(config: &AppConfig, access_code_hash: String) -> Result<Router, AppError> {
     let auth_config = WebAuthConfig::new(
@@ -64,12 +50,12 @@ pub fn build_router(config: &AppConfig, access_code_hash: String) -> Result<Rout
     };
 
     let live_status_service = LiveStatusService::new();
-    let channel_state = ChannelState {
+    let channel_state = routes::ChannelState {
         live_status: live_status_service.clone(),
     };
 
-    let live_status_state = LiveStatusState {
-        service: live_status_service,
+    let live_status_state = routes::LiveStatusState {
+        service: live_status_service.clone(),
         catalog: catalog_service.clone(),
     };
 
@@ -113,94 +99,20 @@ pub fn build_router(config: &AppConfig, access_code_hash: String) -> Result<Rout
 
     let stream_proxy_state = stream_proxy::StreamProxyState::new(stream_service.clone());
 
-    let recording_state = RecordingState {
+    let recording_state = routes::RecordingState {
         service: recording_service,
         default_quality: config.recording.default_quality.clone(),
     };
 
+    // Build route modules
     let channel_routes = routes::channel_routes(channel_state, auth_config.clone());
-
     let live_status_routes = routes::live_status_routes(live_status_state, auth_config.clone());
-
-    let protected_routes = Router::new()
-        .route("/api/channels", get(list_channels))
-        .route("/api/watch-ticket", post(create_watch_ticket))
-        .route("/api/quality-switch", get(quality_switch_handler))
-        .route("/watch/{ticket}", get(render_watch_page))
-        .with_state(protected_state)
-        .layer(middleware::from_fn_with_state(
-            auth_config.clone(),
-            auth::require_session_middleware,
-        ));
-
-    let recording_routes = Router::new()
-        .route("/api/recordings/start", post(start_recording))
-        .route("/api/recordings/stop", post(stop_recording))
-        .route("/api/recordings/pin", post(pin_recording_file))
-        .route("/api/recordings/unpin", post(unpin_recording_file))
-        .route("/api/recordings/delete", post(delete_recording_file))
-        .route("/api/recordings/playback-file", get(play_recording_asset))
-        .route("/api/recordings/hls-playlist", get(serve_hls_playlist))
-        .route("/api/recordings", get(get_recordings))
-        .route("/api/recording-rules", get(get_recording_rules))
-        .route("/api/recording-rules", post(upsert_recording_rule))
-        .route(
-            "/api/recording-rules/{channel_login}",
-            delete(delete_recording_rule),
-        )
-        .with_state(recording_state)
-        .layer(middleware::from_fn_with_state(
-            auth_config.clone(),
-            auth::require_session_middleware,
-        ));
-
-    let twitch_routes = Router::new()
-        .route("/api/twitch/status", get(twitch_auth::get_status))
-        .route("/api/twitch/connect", get(twitch_auth::connect))
-        .route("/api/twitch/callback", get(twitch_auth::callback))
-        .route("/api/twitch/disconnect", post(twitch_auth::disconnect))
-        .with_state(twitch_state)
-        .layer(middleware::from_fn_with_state(
-            auth_config.clone(),
-            auth::require_session_middleware,
-        ));
-
-    let chat_routes = Router::new()
-        .route("/api/chat/status", get(chat::status))
-        .route("/api/chat/emotes", get(chat::emotes))
-        .route("/api/chat/subscribe", post(chat::subscribe))
-        .route("/api/chat/subscribe/{login}", delete(chat::unsubscribe))
-        .route("/api/chat/events/{login}", get(chat::events))
-        .route("/api/chat/send", post(chat::send))
-        .with_state(chat_state)
-        .layer(middleware::from_fn_with_state(
-            auth_config.clone(),
-            auth::require_session_middleware,
-        ));
-
-    let stream_routes = Router::new()
-        .route(
-            "/stream/{stream_id}/{session_token}/manifest",
-            get(stream_proxy::proxy_manifest),
-        )
-        .route(
-            "/stream/{stream_id}/{session_token}/manifest/{quality}",
-            get(stream_proxy::proxy_variant_manifest),
-        )
-        .route(
-            "/stream/{stream_id}/{session_token}/{quality}/{*segment}",
-            get(stream_proxy::proxy_segment),
-        )
-        .with_state(stream_proxy_state);
-
-    let auth_routes = Router::new()
-        .route("/auth/login", post(auth::login))
-        .route("/auth/logout", post(auth::logout))
-        .route("/auth/session", get(auth::session_status))
-        .route("/auth/qr/create", get(auth::create_qr_session))
-        .route("/auth/qr/status/{token}", get(auth::qr_status))
-        .route("/auth/qr/claim/{token}", post(auth::qr_claim))
-        .with_state(auth_config.clone());
+    let watch_routes = routes::watch_routes(protected_state, auth_config.clone());
+    let recording_routes = routes::recording_routes(recording_state, auth_config.clone());
+    let twitch_routes = routes::twitch_routes(twitch_state, auth_config.clone());
+    let chat_routes = routes::chat_routes(chat_state, auth_config.clone());
+    let stream_routes = routes::stream_routes(stream_proxy_state);
+    let auth_routes = routes::auth_routes(auth_config.clone());
 
     let base_path = std::env::current_dir().unwrap_or_else(|_| PathBuf::from("."));
 
@@ -209,7 +121,7 @@ pub fn build_router(config: &AppConfig, access_code_hash: String) -> Result<Rout
 
     let images_path = channels::images_dir().unwrap_or_else(|| PathBuf::from("/tmp"));
     let youtube_images_path =
-        youtube_channels::images_dir().unwrap_or_else(|| PathBuf::from("/tmp"));
+        crate::youtube_channels::images_dir().unwrap_or_else(|| PathBuf::from("/tmp"));
 
     let youtube_routes = youtube::build_routes(auth_config.clone(), config);
 
@@ -218,7 +130,7 @@ pub fn build_router(config: &AppConfig, access_code_hash: String) -> Result<Rout
         .merge(auth_routes)
         .merge(channel_routes)
         .merge(live_status_routes)
-        .merge(protected_routes)
+        .merge(watch_routes)
         .merge(recording_routes)
         .merge(twitch_routes)
         .merge(chat_routes)
@@ -237,940 +149,7 @@ pub fn build_router(config: &AppConfig, access_code_hash: String) -> Result<Rout
     Ok(router)
 }
 
-#[derive(Debug, Serialize)]
-struct ChannelsResponse {
-    channels: Vec<ChannelItem>,
-}
-
-#[derive(Debug, Serialize)]
-struct ChannelItem {
-    login: String,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    image_url: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    display_name: Option<String>,
-    source: String,
-    removable: bool,
-}
-
-#[derive(Debug, Deserialize)]
-struct WatchTicketRequest {
-    channel_login: String,
-}
-
-#[derive(Debug, Serialize)]
-struct WatchTicketResponse {
-    watch_url: String,
-}
-
-#[derive(Debug, Clone)]
-struct ProtectedState {
-    auth: WebAuthConfig,
-    playback: PlaybackTicketService,
-    stream: stream_proxy::StreamSessionService,
-    catalog: ChannelCatalogService,
-}
-
-#[derive(Debug, Clone)]
-struct RecordingState {
-    service: RecordingService,
-    default_quality: String,
-}
-
-#[derive(Debug, Deserialize)]
-struct StartRecordingRequest {
-    channel_login: String,
-    #[serde(default)]
-    quality: Option<String>,
-    #[serde(default)]
-    stream_title: Option<String>,
-}
-
-#[derive(Debug, Deserialize)]
-struct StopRecordingRequest {
-    channel_login: String,
-}
-
-#[derive(Debug, Deserialize)]
-struct UpsertRecordingRuleRequest {
-    channel_login: String,
-    enabled: bool,
-    #[serde(default)]
-    quality: Option<String>,
-    #[serde(default)]
-    stop_when_offline: Option<bool>,
-    #[serde(default)]
-    max_duration_minutes: Option<u64>,
-    #[serde(default)]
-    keep_last_videos: Option<u64>,
-}
-
-#[derive(Debug, Deserialize)]
-struct DeleteRecordingFileRequest {
-    bucket: String,
-    channel_login: String,
-    filename: String,
-}
-
-#[derive(Debug, Deserialize)]
-struct PinRecordingFileRequest {
-    bucket: String,
-    channel_login: String,
-    filename: String,
-}
-
-#[derive(Debug, Deserialize)]
-struct PlayRecordingAssetQuery {
-    channel_login: String,
-    filename: String,
-}
-
-#[derive(Debug, Serialize)]
-struct RecordingRulesResponse {
-    rules: Vec<RecordingRule>,
-}
-
-#[derive(Debug, Serialize)]
-struct RecordingsResponse {
-    active: Vec<ActiveRecording>,
-    completed: Vec<crate::recording::RecordingFileEntry>,
-    incomplete: Vec<crate::recording::RecordingFileEntry>,
-}
-
-async fn start_recording(
-    State(state): State<RecordingState>,
-    Json(payload): Json<StartRecordingRequest>,
-) -> Response {
-    let quality = payload
-        .quality
-        .unwrap_or_else(|| state.default_quality.clone());
-    let quality = match RecordingService::validate_quality(&quality) {
-        Ok(value) => value,
-        Err(_) => return error_response(StatusCode::BAD_REQUEST, "invalid quality"),
-    };
-
-    match state
-        .service
-        .start_recording(
-            &payload.channel_login,
-            &quality,
-            RecordingMode::Manual,
-            payload.stream_title.as_deref(),
-        )
-        .await
-    {
-        Ok(active) => (StatusCode::OK, Json(active)).into_response(),
-        Err(error) => {
-            let (status, message) = classify_recording_error(&error);
-            if status == StatusCode::INTERNAL_SERVER_ERROR {
-                tracing::error!(error = %error, "manual recording start failed");
-            }
-            error_response(status, message)
-        }
-    }
-}
-
-async fn stop_recording(
-    State(state): State<RecordingState>,
-    Json(payload): Json<StopRecordingRequest>,
-) -> Response {
-    match state.service.stop_recording(&payload.channel_login).await {
-        Ok(active) => (StatusCode::OK, Json(active)).into_response(),
-        Err(error) => {
-            let (status, message) = classify_recording_error(&error);
-            if status == StatusCode::INTERNAL_SERVER_ERROR {
-                tracing::error!(error = %error, "recording stop failed");
-            }
-            error_response(status, message)
-        }
-    }
-}
-
-async fn get_recordings(State(state): State<RecordingState>) -> Json<RecordingsResponse> {
-    let overview = state.service.list_overview(15).await;
-    Json(RecordingsResponse {
-        active: overview.active,
-        completed: overview.completed,
-        incomplete: overview.incomplete,
-    })
-}
-
-async fn delete_recording_file(
-    State(state): State<RecordingState>,
-    Json(payload): Json<DeleteRecordingFileRequest>,
-) -> Response {
-    let bucket = match payload.bucket.as_str() {
-        "completed" => RecordingBucket::Completed,
-        "incomplete" => RecordingBucket::Incomplete,
-        _ => return error_response(StatusCode::BAD_REQUEST, "invalid recording bucket"),
-    };
-
-    match state
-        .service
-        .delete_recording_file(bucket, &payload.channel_login, &payload.filename)
-    {
-        Ok(()) => StatusCode::NO_CONTENT.into_response(),
-        Err(error) => {
-            let (status, message) = classify_recording_error(&error);
-            if status == StatusCode::INTERNAL_SERVER_ERROR {
-                tracing::error!(error = %error, "recording file delete failed");
-            }
-            error_response(status, message)
-        }
-    }
-}
-
-async fn pin_recording_file(
-    State(state): State<RecordingState>,
-    Json(payload): Json<PinRecordingFileRequest>,
-) -> Response {
-    if payload.bucket != "completed" {
-        return error_response(
-            StatusCode::BAD_REQUEST,
-            "pinning is only supported for completed recordings",
-        );
-    }
-
-    match state
-        .service
-        .pin_recording_file(&payload.channel_login, &payload.filename)
-    {
-        Ok(()) => StatusCode::NO_CONTENT.into_response(),
-        Err(error) => {
-            let (status, message) = classify_recording_error(&error);
-            if status == StatusCode::INTERNAL_SERVER_ERROR {
-                tracing::error!(error = %error, "recording file pin failed");
-            }
-            error_response(status, message)
-        }
-    }
-}
-
-async fn unpin_recording_file(
-    State(state): State<RecordingState>,
-    Json(payload): Json<PinRecordingFileRequest>,
-) -> Response {
-    if payload.bucket != "completed" {
-        return error_response(
-            StatusCode::BAD_REQUEST,
-            "unpinning is only supported for completed recordings",
-        );
-    }
-
-    match state
-        .service
-        .unpin_recording_file(&payload.channel_login, &payload.filename)
-    {
-        Ok(()) => StatusCode::NO_CONTENT.into_response(),
-        Err(error) => {
-            let (status, message) = classify_recording_error(&error);
-            if status == StatusCode::INTERNAL_SERVER_ERROR {
-                tracing::error!(error = %error, "recording file unpin failed");
-            }
-            error_response(status, message)
-        }
-    }
-}
-
-async fn play_recording_asset(
-    State(state): State<RecordingState>,
-    Query(query): Query<PlayRecordingAssetQuery>,
-    headers: HeaderMap,
-) -> Response {
-    let media_path = match state
-        .service
-        .resolve_completed_file_path(&query.channel_login, &query.filename)
-    {
-        Ok(path) => path,
-        Err(error) => {
-            let (status, message) = classify_recording_error(&error);
-            return error_response(status, message);
-        }
-    };
-
-    if !media_path.exists() {
-        return error_response(StatusCode::NOT_FOUND, "recording playback asset not found");
-    }
-
-    let file_size = match tokio::fs::metadata(&media_path).await {
-        Ok(meta) => meta.len(),
-        Err(error) => {
-            tracing::error!(error = %error, path = %media_path.display(), "failed to read playback media metadata");
-            return error_response(
-                StatusCode::INTERNAL_SERVER_ERROR,
-                "recording playback failed",
-            );
-        }
-    };
-
-    if let Some(range_header) = headers.get(header::RANGE) {
-        let Ok(range_str) = range_header.to_str() else {
-            return error_response(StatusCode::BAD_REQUEST, "invalid range header");
-        };
-        let Some(range_spec) = range_str.strip_prefix("bytes=") else {
-            return error_response(StatusCode::BAD_REQUEST, "invalid range header");
-        };
-        let Some((start_str, end_str)) = range_spec.split_once('-') else {
-            return error_response(StatusCode::BAD_REQUEST, "invalid range header");
-        };
-
-        let start: u64 = match start_str.parse() {
-            Ok(v) => v,
-            Err(_) => return error_response(StatusCode::BAD_REQUEST, "invalid range start"),
-        };
-
-        // HLS uses exact byte ranges from the m3u8 - open-ended ranges not supported
-        let end: u64 = if end_str.is_empty() {
-            return error_response(StatusCode::BAD_REQUEST, "open-ended ranges not supported");
-        } else {
-            match end_str.parse() {
-                Ok(v) => v,
-                Err(_) => return error_response(StatusCode::BAD_REQUEST, "invalid range end"),
-            }
-        };
-
-        if start >= file_size || end >= file_size || end < start {
-            return error_response(StatusCode::RANGE_NOT_SATISFIABLE, "range not satisfiable");
-        }
-
-        let length = end - start + 1;
-        let media_stream = match stream_file_range(&media_path, start, length).await {
-            Ok(stream) => stream,
-            Err(error) => {
-                tracing::error!(error = %error, path = %media_path.display(), "failed to read playback range");
-                return error_response(
-                    StatusCode::INTERNAL_SERVER_ERROR,
-                    "recording playback failed",
-                );
-            }
-        };
-
-        let content_range = format!("bytes {start}-{end}/{file_size}");
-        let mut response = Response::new(Body::from_stream(media_stream));
-        *response.status_mut() = StatusCode::PARTIAL_CONTENT;
-        response
-            .headers_mut()
-            .insert(header::CONTENT_TYPE, HeaderValue::from_static("video/mp4"));
-        response.headers_mut().insert(
-            header::CONTENT_RANGE,
-            HeaderValue::from_str(&content_range).unwrap(),
-        );
-        response.headers_mut().insert(
-            header::CONTENT_LENGTH,
-            HeaderValue::from_str(&length.to_string()).unwrap(),
-        );
-        response
-            .headers_mut()
-            .insert(header::ACCEPT_RANGES, HeaderValue::from_static("bytes"));
-        response.headers_mut().insert(
-            header::CACHE_CONTROL,
-            HeaderValue::from_static("public, max-age=2592000, immutable"),
-        );
-        response
-    } else {
-        let media_stream = match stream_file_range(&media_path, 0, file_size).await {
-            Ok(stream) => stream,
-            Err(error) => {
-                tracing::error!(error = %error, path = %media_path.display(), "failed to read playback media");
-                return error_response(
-                    StatusCode::INTERNAL_SERVER_ERROR,
-                    "recording playback failed",
-                );
-            }
-        };
-        let mut response = Response::new(Body::from_stream(media_stream));
-        *response.status_mut() = StatusCode::OK;
-        response
-            .headers_mut()
-            .insert(header::CONTENT_TYPE, HeaderValue::from_static("video/mp4"));
-        response.headers_mut().insert(
-            header::CONTENT_LENGTH,
-            HeaderValue::from_str(&file_size.to_string()).unwrap(),
-        );
-        response
-            .headers_mut()
-            .insert(header::ACCEPT_RANGES, HeaderValue::from_static("bytes"));
-        response.headers_mut().insert(
-            header::CACHE_CONTROL,
-            HeaderValue::from_static("public, max-age=2592000, immutable"),
-        );
-        response
-    }
-}
-
-#[derive(Deserialize)]
-struct ServeHlsPlaylistQuery {
-    channel_login: String,
-    filename: String,
-}
-
-async fn serve_hls_playlist(
-    State(state): State<RecordingState>,
-    Query(query): Query<ServeHlsPlaylistQuery>,
-) -> Response {
-    // Resolve the MP4 path
-    let mp4_path = match state
-        .service
-        .resolve_completed_file_path(&query.channel_login, &query.filename)
-    {
-        Ok(path) => path,
-        Err(error) => {
-            let (status, message) = classify_recording_error(&error);
-            return error_response(status, message);
-        }
-    };
-
-    if !mp4_path.exists() {
-        return error_response(StatusCode::NOT_FOUND, "recording not found");
-    }
-
-    // Look for the .m3u8 playlist file
-    let playlist_path = mp4_path.with_extension("m3u8");
-    if !playlist_path.exists() {
-        return error_response(StatusCode::NOT_FOUND, "hls playlist not found");
-    }
-
-    // Read and serve the playlist
-    let playlist_content = match tokio::fs::read_to_string(&playlist_path).await {
-        Ok(content) => content,
-        Err(error) => {
-            tracing::error!(error = %error, path = %playlist_path.display(), "failed to read hls playlist");
-            return error_response(StatusCode::INTERNAL_SERVER_ERROR, "failed to read playlist");
-        }
-    };
-
-    let mut response = Response::new(Body::from(playlist_content));
-    *response.status_mut() = StatusCode::OK;
-    response.headers_mut().insert(
-        header::CONTENT_TYPE,
-        HeaderValue::from_static("application/vnd.apple.mpegurl"),
-    );
-    response.headers_mut().insert(
-        header::CACHE_CONTROL,
-        HeaderValue::from_static("no-cache, must-revalidate"),
-    );
-    response
-}
-
-async fn stream_file_range(
-    path: &std::path::Path,
-    start: u64,
-    length: u64,
-) -> Result<
-    std::pin::Pin<Box<dyn futures_util::Stream<Item = Result<Bytes, std::io::Error>> + Send>>,
-    String,
-> {
-    use tokio::io::{AsyncReadExt, AsyncSeekExt};
-
-    let mut file = tokio::fs::File::open(path)
-        .await
-        .map_err(|error| format!("failed to open playback media: {error}"))?;
-    file.seek(std::io::SeekFrom::Start(start))
-        .await
-        .map_err(|error| format!("failed to seek playback media: {error}"))?;
-    let stream = stream::try_unfold((file, length), |(mut file, remaining)| async move {
-        const CHUNK_SIZE: usize = 4 * 1024 * 1024; // 4 MiB chunks for HDD optimization
-
-        if remaining == 0 {
-            return Ok(None);
-        }
-
-        let next_len = usize::try_from(remaining.min(CHUNK_SIZE as u64)).unwrap_or(CHUNK_SIZE);
-        let mut chunk = vec![0_u8; next_len];
-        let read = file.read(&mut chunk).await?;
-        if read == 0 {
-            return Ok(None);
-        }
-        chunk.truncate(read);
-        let read_u64 = u64::try_from(read).unwrap_or(0);
-        let next_remaining = remaining.saturating_sub(read_u64);
-        Ok(Some((Bytes::from(chunk), (file, next_remaining))))
-    });
-
-    Ok(Box::pin(stream))
-}
-
-async fn get_recording_rules() -> Response {
-    match recording_rules::load_rules() {
-        Ok(rules) => (StatusCode::OK, Json(RecordingRulesResponse { rules })).into_response(),
-        Err(error) => {
-            tracing::error!(error = %error, "recording rules load failed");
-            error_response(
-                StatusCode::INTERNAL_SERVER_ERROR,
-                "recording rules file read/write failure",
-            )
-        }
-    }
-}
-
-async fn upsert_recording_rule(
-    State(state): State<RecordingState>,
-    Json(payload): Json<UpsertRecordingRuleRequest>,
-) -> Response {
-    let channel_login = payload.channel_login.trim().to_ascii_lowercase();
-    if channel_login.is_empty() {
-        return error_response(StatusCode::BAD_REQUEST, "channel login cannot be empty");
-    }
-
-    let quality_input = payload
-        .quality
-        .unwrap_or_else(|| state.default_quality.clone());
-    let quality = match RecordingService::validate_quality(&quality_input) {
-        Ok(value) => value,
-        Err(_) => return error_response(StatusCode::BAD_REQUEST, "invalid quality"),
-    };
-
-    if payload.keep_last_videos == Some(0) {
-        return error_response(StatusCode::BAD_REQUEST, "keep_last_videos must be >= 1");
-    }
-
-    let rule = RecordingRule {
-        channel_login,
-        enabled: payload.enabled,
-        quality,
-        stop_when_offline: payload.stop_when_offline.unwrap_or(true),
-        max_duration_minutes: payload.max_duration_minutes,
-        keep_last_videos: payload.keep_last_videos,
-    };
-
-    match recording_rules::upsert_rule(rule) {
-        Ok(saved) => (StatusCode::OK, Json(saved)).into_response(),
-        Err(error) => {
-            tracing::error!(error = %error, "recording rule save failed");
-            error_response(
-                StatusCode::INTERNAL_SERVER_ERROR,
-                "recording rules file read/write failure",
-            )
-        }
-    }
-}
-
-async fn delete_recording_rule(Path(channel_login): Path<String>) -> Response {
-    let normalized = channel_login.trim().to_ascii_lowercase();
-    if normalized.is_empty() {
-        return error_response(StatusCode::BAD_REQUEST, "channel login cannot be empty");
-    }
-
-    match recording_rules::delete_rule(&normalized) {
-        Ok(true) => StatusCode::NO_CONTENT.into_response(),
-        Ok(false) => error_response(StatusCode::NOT_FOUND, "recording rule not found"),
-        Err(error) => {
-            tracing::error!(error = %error, "recording rule delete failed");
-            error_response(
-                StatusCode::INTERNAL_SERVER_ERROR,
-                "recording rules file read/write failure",
-            )
-        }
-    }
-}
-
-fn classify_recording_error(error: &RecordingError) -> (StatusCode, &'static str) {
-    match error {
-        RecordingError::EmptyChannelLogin => {
-            (StatusCode::BAD_REQUEST, "channel login cannot be empty")
-        }
-        RecordingError::InvalidQuality => (StatusCode::BAD_REQUEST, "invalid quality"),
-        RecordingError::AlreadyActive => (StatusCode::CONFLICT, "recording already active"),
-        RecordingError::NotActive => (StatusCode::NOT_FOUND, "recording not active"),
-        RecordingError::FileNotFound => (StatusCode::NOT_FOUND, "recording file not found"),
-        RecordingError::EmptyFilename => (StatusCode::BAD_REQUEST, "filename cannot be empty"),
-        RecordingError::InvalidFilename => (StatusCode::BAD_REQUEST, "invalid filename"),
-        RecordingError::DeleteFailed(_) => {
-            (StatusCode::INTERNAL_SERVER_ERROR, "recording delete failed")
-        }
-        RecordingError::SpawnFailed(_) => (StatusCode::BAD_GATEWAY, "streamlink spawn failed"),
-        RecordingError::DirectoryNotWritable(_) => (
-            StatusCode::INTERNAL_SERVER_ERROR,
-            "recordings directory not writable",
-        ),
-        RecordingError::PinFailed(_) | RecordingError::UnpinFailed(_) | RecordingError::Io(_) => (
-            StatusCode::INTERNAL_SERVER_ERROR,
-            "recording operation failed",
-        ),
-    }
-}
-
-async fn list_channels(State(state): State<ProtectedState>) -> Json<ChannelsResponse> {
-    let mut channels_list: Vec<ChannelItem> = state
-        .catalog
-        .list_channels()
-        .await
-        .into_iter()
-        .map(channel_item_from_catalog)
-        .collect();
-
-    channels_list.sort_by_key(|c| c.login.to_lowercase());
-
-    Json(ChannelsResponse {
-        channels: channels_list,
-    })
-}
-
-fn channel_item_from_catalog(item: CatalogChannel) -> ChannelItem {
-    let source = match item.source {
-        crate::channel_catalog::ChannelSource::Manual => "manual",
-        crate::channel_catalog::ChannelSource::Followed => "followed",
-        crate::channel_catalog::ChannelSource::Both => "both",
-    };
-
-    ChannelItem {
-        login: item.login,
-        image_url: item.image_url,
-        display_name: item.display_name,
-        source: source.to_string(),
-        removable: item.removable,
-    }
-}
-
-async fn create_watch_ticket(
-    State(state): State<ProtectedState>,
-    headers: HeaderMap,
-    Json(payload): Json<WatchTicketRequest>,
-) -> Response {
-    if !state.catalog.has_channel(&payload.channel_login).await {
-        return error_response(StatusCode::BAD_REQUEST, "channel is not in channel list");
-    }
-
-    let Some(session_token) = state.auth.session_token_from_headers(&headers) else {
-        return error_response(StatusCode::UNAUTHORIZED, "authentication required");
-    };
-
-    match state
-        .playback
-        .issue_ticket(&session_token, &payload.channel_login)
-    {
-        Ok(ticket) => {
-            let response = WatchTicketResponse {
-                watch_url: format!("/watch/{ticket}"),
-            };
-            (StatusCode::OK, Json(response)).into_response()
-        }
-        Err(_) => error_response(
-            StatusCode::INTERNAL_SERVER_ERROR,
-            "failed to issue watch ticket",
-        ),
-    }
-}
-
-#[derive(Debug, Deserialize)]
-struct QualitySwitchQuery {
-    channel_login: String,
-    quality: String,
-}
-
-#[derive(Debug, Serialize)]
-struct QualitySwitchResponse {
-    watch_url: String,
-    quality: String,
-}
-
-async fn quality_switch_handler(
-    State(state): State<ProtectedState>,
-    headers: HeaderMap,
-    Query(query): Query<QualitySwitchQuery>,
-) -> Response {
-    if !state.catalog.has_channel(&query.channel_login).await {
-        return error_response(StatusCode::BAD_REQUEST, "channel is not in channel list");
-    }
-
-    let Some(session_token) = state.auth.session_token_from_headers(&headers) else {
-        return error_response(StatusCode::UNAUTHORIZED, "authentication required");
-    };
-
-    match state
-        .playback
-        .issue_ticket(&session_token, &query.channel_login)
-    {
-        Ok(ticket) => {
-            let stream_id = &ticket;
-
-            if let Err(e) = state
-                .stream
-                .open_session(
-                    stream_id,
-                    &query.channel_login,
-                    &session_token,
-                    &query.quality,
-                )
-                .await
-            {
-                tracing::error!(error = ?e, channel = %query.channel_login, quality = %query.quality, "failed to open stream session for quality switch");
-                return error_response(
-                    StatusCode::BAD_GATEWAY,
-                    "failed to open stream with requested quality",
-                );
-            }
-
-            let response = QualitySwitchResponse {
-                watch_url: format!("/watch/{ticket}"),
-                quality: query.quality,
-            };
-            (StatusCode::OK, Json(response)).into_response()
-        }
-        Err(_) => error_response(
-            StatusCode::INTERNAL_SERVER_ERROR,
-            "failed to issue watch ticket",
-        ),
-    }
-}
-
-async fn render_watch_page(
-    State(state): State<ProtectedState>,
-    headers: HeaderMap,
-    Path(ticket): Path<String>,
-    Query(query): Query<stream_proxy::RelayQuery>,
-) -> Response {
-    let Some(session_token) = state.auth.session_token_from_headers(&headers) else {
-        return error_response(StatusCode::UNAUTHORIZED, "authentication required");
-    };
-
-    let validated = match state.playback.validate_ticket(&ticket, &session_token) {
-        Ok(v) => v,
-        Err(PlaybackTicketError::InvalidTicket) | Err(PlaybackTicketError::ExpiredTicket) => {
-            return error_response(StatusCode::UNAUTHORIZED, "invalid or expired watch ticket");
-        }
-        Err(PlaybackTicketError::SessionMismatch) => {
-            return error_response(
-                StatusCode::FORBIDDEN,
-                "watch ticket belongs to a different session",
-            );
-        }
-    };
-
-    if let Err(e) = state
-        .stream
-        .open_session(&ticket, &validated.channel_login, &session_token, "best")
-        .await
-    {
-        return match e {
-            stream_proxy::StreamError::HlsFetchFailed(msg) => {
-                tracing::error!(error = %msg, channel = %validated.channel_login, "failed to open stream session");
-                render_error_page(
-                    &validated.channel_login,
-                    "Stream unavailable. The channel may be offline or not accessible.",
-                )
-            }
-            _ => error_response(
-                StatusCode::INTERNAL_SERVER_ERROR,
-                "failed to open stream session",
-            ),
-        };
-    }
-
-    let html = render_stream_page(
-        &validated.channel_login,
-        &ticket,
-        &session_token,
-        query.force_relay(),
-    );
-
-    Html(html).into_response()
-}
-
-fn render_stream_page(
-    channel: &str,
-    stream_id: &str,
-    session_token: &str,
-    force_relay: bool,
-) -> String {
-    let relay_suffix = if force_relay { "?relay=1" } else { "" };
-    let manifest_url = format!("/stream/{stream_id}/{session_token}/manifest{relay_suffix}");
-    let template = include_str!("templates/watch.html");
-    let mut watch_config = serde_json::json!({
-        "channel": channel,
-        "manifestUrl": manifest_url,
-        "relay": force_relay,
-    })
-    .to_string();
-    watch_config = watch_config.replace("</script>", "<\\/script>");
-    let bootstrap = format!("window.__WATCH_CONFIG__ = {watch_config};");
-
-    template
-        .replace("__CHANNEL__", channel)
-        .replace("__APP_VERSION__", APP_VERSION)
-        .replace("__WATCH_BOOTSTRAP__", &bootstrap)
-}
-
-fn render_error_page(channel: &str, message: &str) -> Response {
-    let html = format!(
-        r#"<!doctype html>
-<html>
-<head>
-<meta charset="utf-8">
-<meta name="viewport" content="width=device-width, initial-scale=1">
-<title>Watch {channel}</title>
-<style>
-  /* Tokyo Night Moon theme tokens */
-  :root {{
-    --bg: #1e2030;
-    --bg-soft: #222436;
-    --surface: #2f334d;
-    --fg: #c8d3f5;
-    --muted: #a9b8e8;
-    --border: #444a73;
-  }}
-  * {{ margin: 0; padding: 0; box-sizing: border-box; }}
-  body {{
-    background: var(--bg);
-    color: var(--fg);
-    font-family: system-ui, -apple-system, 'Segoe UI', sans-serif;
-    min-height: 100vh;
-    display: flex;
-    flex-direction: column;
-  }}
-  header {{
-    padding: 0.75rem 1rem;
-    border-bottom: 1px solid var(--border);
-    display: flex;
-    align-items: center;
-    gap: 0.75rem;
-  }}
-  header strong {{ font-size: 1rem; font-weight: 700; text-transform: lowercase; }}
-  .error-screen {{
-    flex: 1;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    padding: 2rem;
-  }}
-  .error-box {{
-    text-align: center;
-    max-width: 28rem;
-    background: rgba(47, 51, 77, 0.95);
-    border: 1px solid color-mix(in srgb, var(--border) 65%, transparent);
-    border-radius: 1rem;
-    padding: 1.5rem;
-  }}
-  .error-box p {{
-    color: var(--muted);
-    line-height: 1.6;
-  }}
-</style>
-</head>
-<body>
-<header>
-  <strong>{channel}</strong>
-  <span>via Twitch Relay · v{version}</span>
-</header>
-<div class="error-screen">
-  <div class="error-box">
-    <p>{message}</p>
-  </div>
-</div>
-</body>
-</html>"#,
-        channel = channel,
-        message = message,
-        version = APP_VERSION
-    );
-
-    (StatusCode::OK, Html(html)).into_response()
-}
-
-fn error_response(status: StatusCode, message: &str) -> Response {
-    (status, Json(serde_json::json!({ "error": message }))).into_response()
-}
-
 #[cfg(test)]
 mod tests {
-    use super::*;
-    use crate::recording::RecordingError;
-
-    #[test]
-    fn classify_recording_error_maps_empty_channel_login() {
-        let (status, message) = classify_recording_error(&RecordingError::EmptyChannelLogin);
-        assert_eq!(status, StatusCode::BAD_REQUEST);
-        assert_eq!(message, "channel login cannot be empty");
-    }
-
-    #[test]
-    fn classify_recording_error_maps_invalid_quality() {
-        let (status, message) = classify_recording_error(&RecordingError::InvalidQuality);
-        assert_eq!(status, StatusCode::BAD_REQUEST);
-        assert_eq!(message, "invalid quality");
-    }
-
-    #[test]
-    fn classify_recording_error_maps_already_active() {
-        let (status, message) = classify_recording_error(&RecordingError::AlreadyActive);
-        assert_eq!(status, StatusCode::CONFLICT);
-        assert_eq!(message, "recording already active");
-    }
-
-    #[test]
-    fn classify_recording_error_maps_not_active() {
-        let (status, message) = classify_recording_error(&RecordingError::NotActive);
-        assert_eq!(status, StatusCode::NOT_FOUND);
-        assert_eq!(message, "recording not active");
-    }
-
-    #[test]
-    fn classify_recording_error_maps_file_not_found() {
-        let (status, message) = classify_recording_error(&RecordingError::FileNotFound);
-        assert_eq!(status, StatusCode::NOT_FOUND);
-        assert_eq!(message, "recording file not found");
-    }
-
-    #[test]
-    fn classify_recording_error_maps_empty_filename() {
-        let (status, message) = classify_recording_error(&RecordingError::EmptyFilename);
-        assert_eq!(status, StatusCode::BAD_REQUEST);
-        assert_eq!(message, "filename cannot be empty");
-    }
-
-    #[test]
-    fn classify_recording_error_maps_invalid_filename() {
-        let (status, message) = classify_recording_error(&RecordingError::InvalidFilename);
-        assert_eq!(status, StatusCode::BAD_REQUEST);
-        assert_eq!(message, "invalid filename");
-    }
-
-    #[test]
-    fn classify_recording_error_maps_delete_failed() {
-        let (status, message) =
-            classify_recording_error(&RecordingError::DeleteFailed("some error".to_string()));
-        assert_eq!(status, StatusCode::INTERNAL_SERVER_ERROR);
-        assert_eq!(message, "recording delete failed");
-    }
-
-    #[test]
-    fn classify_recording_error_maps_spawn_failed() {
-        let (status, message) =
-            classify_recording_error(&RecordingError::SpawnFailed("some error".to_string()));
-        assert_eq!(status, StatusCode::BAD_GATEWAY);
-        assert_eq!(message, "streamlink spawn failed");
-    }
-
-    #[test]
-    fn classify_recording_error_maps_directory_not_writable() {
-        let (status, message) = classify_recording_error(&RecordingError::DirectoryNotWritable(
-            "some error".to_string(),
-        ));
-        assert_eq!(status, StatusCode::INTERNAL_SERVER_ERROR);
-        assert_eq!(message, "recordings directory not writable");
-    }
-
-    #[test]
-    fn classify_recording_error_maps_pin_failed_to_fallback() {
-        let (status, message) =
-            classify_recording_error(&RecordingError::PinFailed("some error".to_string()));
-        assert_eq!(status, StatusCode::INTERNAL_SERVER_ERROR);
-        assert_eq!(message, "recording operation failed");
-    }
-
-    #[test]
-    fn classify_recording_error_maps_unpin_failed_to_fallback() {
-        let (status, message) =
-            classify_recording_error(&RecordingError::UnpinFailed("some error".to_string()));
-        assert_eq!(status, StatusCode::INTERNAL_SERVER_ERROR);
-        assert_eq!(message, "recording operation failed");
-    }
-
-    #[test]
-    fn classify_recording_error_maps_io_to_fallback() {
-        let (status, message) =
-            classify_recording_error(&RecordingError::Io("some error".to_string()));
-        assert_eq!(status, StatusCode::INTERNAL_SERVER_ERROR);
-        assert_eq!(message, "recording operation failed");
-    }
+    // Recording error tests moved to src/routes/recordings.rs
 }
